Skeleton on Crawler class and changed namespaces pga Crawler waas used in namespaces also.
