+++ LTU.SearchEngine.Backend/LTU.SearchEngine.Test/Crawling.Tests/CrawlerTest.cs
@@ -1,0 +1,12 @@
+﻿using LTU.SearchEngine.Infrastructure.Crawling;
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace LTU.SearchEngine.Test.Crawling.Tests
+{
+    public class CrawlerTest
+    {
+        private readonly Crawler _sut;
+    }
+}
