+++ LTU.SearchEngine.Backend/LTU.SearchEngine.Test/Crawling.Tests/ParserTests.cs
@@ -1,7 +1,6 @@
-﻿
-using LTU.SearchEngine.Infrastructure;
+﻿using LTU.SearchEngine.Infrastructure;
 
-namespace LTU.SearchEngine.Test.Crawler.Tests;
+namespace LTU.SearchEngine.Test.Crawling.Tests;
 
 public class ParserTests
 {
